backend/training/requirements.txt
# SIPLACE SX2 预测模型后端依赖

# FastAPI Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# 深度学习框架
tensorflow>=2.14.0
keras>=2.14.0

# 生存分析
lifelines>=0.27.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 数据库
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# 模型导出
tf2onnx>=1.15.0
onnx>=1.14.0

# 任务队列（可选）
celery>=5.3.0
redis>=5.0.0

# 工具库
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-json-logger>=2.0.0
